refactor(host): clarify exact ownership boundaries
