+++ docs/architecture.md
@@ -159,7 +159,7 @@ The Station UI in `station/` is a `@station/client` consumer plus a terminal-hos
 - The `station-station-host` daemon (`packages/station-host`) owns PTYs that outlive the UI. Its socket defaults beside the observer socket at `<state_dir>/run/station-host.sock` (override `STATION_HOST_SOCKET_PATH`).
 - Host output compatibility is an optional generic spawn policy selected by integrations and applied before both scrollback retention and attached-client broadcast, so live and warm-reattached views consume one byte stream.
 - Host retains complete transformed output plus every production-geometry transition within its bounded raw replay budget. Attach returns those ordered events while complete; after eviction it prefers restoration VT from xterm's serializer plus the small set of Station-relevant modes that serializer omits. Semantic capture fails closed between xterm parser boundaries, and the client retries that transient state after later output can complete the sequence. At a safe boundary where exact reconstruction is unavailable, Host retains the live sink and returns RIS-prefixed, control-only reset data that restores the boundary-captured interaction modes and a valid active-buffer cursor anchor without historical content; Station applies it before nudging PTY geometry so cursor-relative child repaint remains positioned from the captured frame. Live resize frames remain ordered with later output.
-- Generic Host ensure is compatibility-only: it reuses the current protocol and exact display version, starts after absent or proven-stale evidence, atomically replaces an idle different-display Host, and visibly refuses a busy or incompatible incumbent. Exact artifact ownership is a separate, unversioned convergence composition over the immutable `{ buildVersion, buildIdentity }` pair. Its strict command binds one configured socket, physical incumbent endpoint, complete canonical terminal-lifetime evidence, action, and absolute deadline before I/O. Protocol-v8 `host.list` remains unchanged; `host.recoveryInventory` supplies the immutable Host identity and complete per-PTY handoff facts in canonical `{ terminalTargetId, ptyId, ptyInstanceId }` order. Each identifier is independently unique; `sessionId` is not globally unique. Exact convergence validates the incumbent on one non-reconnecting lifecycle connection, releases only that lifetime, causally admits one direct successor child with `E0 → H0 → lsof-holder → H1 → E1`, adopts exactly the receipt set, and independently verifies the exact successor endpoint and registry. The fixed final two seconds remain reserved for settling an untransferred child. A successful transfer is the sole ownership milestone; later failures never signal the proven successor.
+- Generic Host ensure is compatibility-only: it reuses the current protocol and exact display version, starts after absent or proven-stale evidence, atomically replaces an idle different-display Host, and visibly refuses a busy or incompatible incumbent. Exact artifact ownership is a separate, unversioned convergence composition over the immutable `{ buildVersion, buildIdentity }` pair. Its strict command binds one configured socket, physical incumbent endpoint, complete canonical terminal-lifetime evidence, action, and absolute deadline before I/O. Protocol-v8 `host.list` remains unchanged; `host.recoveryInventory` supplies the immutable Host identity and complete per-PTY handoff facts in canonical `{ terminalTargetId, ptyId, ptyInstanceId }` order. Each identifier is independently unique; `sessionId` is not globally unique. Exact convergence validates the incumbent on one non-reconnecting lifecycle connection, releases only that lifetime, and admits one direct successor only when the socket endpoint and pinned Host health remain unchanged before and after canonical `lsof` holder evidence. It then adopts exactly the receipt set and independently verifies the exact successor endpoint and registry. The fixed final two seconds remain reserved for settling an untransferred child. A successful transfer is the sole ownership milestone; later failures never signal the proven successor.
 - Each physical Host connection receives an opaque lifecycle owner. A successful `beginHandoff` binds pre-complete complete/abort/disconnect authority to that object; another connection cannot complete, abort, or acquire restoration authority. Disconnect restores only its own pre-complete parked registry, while disconnect after completion is inert. Attach still transfers replay and live frames, never PTY file descriptors; ownership moves by parking per-terminal bridges and adopting their control sockets. Adoption acknowledgements and handoff receipts correlate identities but authorize nothing until exact successor evidence confirms them.
 - Native TUI launch keeps the compatibility-only path unless `STATION_HOST_HANDOFF` is exactly `1`. Under that gate, it performs read-only exact inspection and reuses only an exact display/identity pair; a same-display different-identity Host therefore takes idle replacement or eligible live handoff. Noncanonical, unsupported, drifted, ineligible, or incomplete evidence refuses without `host.list`, cold restore, or local-spawn fallback. Canonical convergence returns the exact successor inventory used for warm restoration, preserving PTY ID, `ptyInstanceId`, replay, and live I/O. Standalone `stn host handoff` uses the same composition while retaining its existing output and exit contract; dry-run performs inspection only. Observer selectors remain independent, so UI admission is not an immutable three-process cohort rule. After invoking the native renderer, CLI composition starts one process-local owner-aware update detection and planning check with no persistent cache. It never calls the plan's apply capability; renderer resolution aborts unfinished discovery without awaiting it, and only an already-completed version-changing plan may print after a normal zero-code, unsignaled exit. Popup and fake-dashboard renderers never start this check.
 - Host output may fan out to many attachments, while a Host-issued attachment lease allows at most one controller to write or resize each PTY. Controller grants advance a per-PTY epoch and revoke the former controller before it can mutate again; detach leaves viewers and the live PTY intact without promoting either. Native renderers cache their desired geometry while viewing and reclaim only for user input, then apply the latest geometry before forwarding that input. Bare PTY identity, resize traffic, process identity, and renderer environment never grant mutation authority.
