+++ integrations/terminal/station/src/host/readStationHostEvidence.ts
@@ -1,25 +1,12 @@
 import {
   StationHostEndpointSchema,
+  type StationHostExactEvidence,
   StationHostExactEvidenceSchema,
   StationHostInspectedHealthSchema,
 } from "@station/contracts";
-import {
-  assertHostReusable,
-  type HostHealthResult,
-  isStationHostCompatibilityError,
-  openStationHostLifecycleSession,
-  type StationHostLifecycleSession,
-} from "@station/host";
+import type { StationHostLifecycleSession } from "@station/host";
 import { probeUnixSocket, readUnixSocketHolderPidsAsync } from "@station/protocol";
-import type { z } from "zod";
-import type {
-  ChildProcessLike,
-  SpawnStationHostInput,
-  StationHostCommand,
-} from "./ensureHostRunning.js";
-import { startStationHostProcess } from "./hostProcess.js";
 
-export type StationHostExactEvidence = z.infer<typeof StationHostExactEvidenceSchema>;
 export type StationHostEndpointObservation =
   | { status: "absent" }
   | {
@@ -32,12 +19,6 @@ export type StationHostEndpointProbe = (
   socketPath: string,
   deadlineMs: number,
 ) => Promise<StationHostEndpointObservation>;
-export type CausalStationHostEvidencePorts = {
-  openSession: typeof openStationHostLifecycleSession;
-  probeEndpoint: StationHostEndpointProbe;
-  readHolders(socketPath: string, deadlineMs: number): Promise<readonly number[]>;
-  now(): number;
-};
 
 /**
  * ADAPTER
@@ -109,146 +90,6 @@ export async function readStationHostEvidence(input: {
   });
 }
 
-type CausalCandidate = {
-  endpoint: StationHostExactEvidence["endpoint"];
-  health: HostHealthResult;
-  holders: readonly number[];
-  session: StationHostLifecycleSession;
-};
-
-/** Produces `E0 → H0 → holder → H1 → E1`; only dial/initial-health failures retry. */
-async function readCausalCandidate(
-  input: {
-    socketPath: string;
-    expectedBuildVersion: string;
-    deadlineMs: number;
-    validate?: (session: StationHostLifecycleSession) => Promise<void>;
-  },
-  ports: CausalStationHostEvidencePorts,
-): Promise<CausalCandidate> {
-  while (ports.now() < input.deadlineMs) {
-    const e0 = await ports.probeEndpoint(input.socketPath, input.deadlineMs);
-    if (e0.status === "absent" || e0.status === "stale") {
-      await retryDelay(ports.now, input.deadlineMs);
-      continue;
-    }
-    if (e0.status === "inaccessible") throw e0.error;
-    let session: StationHostLifecycleSession | undefined;
-    let admitted = false;
-    try {
-      session = await ports.openSession({
-        socketPath: input.socketPath,
-        expectedBuildVersion: input.expectedBuildVersion,
-        deadlineMs: input.deadlineMs,
-      });
-      const h0 = await session.health();
-      assertHostReusable(h0, input.expectedBuildVersion);
-      admitted = true;
-      await input.validate?.(session);
-      const holders = await ports.readHolders(input.socketPath, input.deadlineMs);
-      const h1 = await session.health();
-      const e1 = await ports.probeEndpoint(input.socketPath, input.deadlineMs);
-      if (
-        ports.now() >= input.deadlineMs ||
-        e1.status !== "listening" ||
-        !stationHostEndpointsMatch(e0.endpoint, e1.endpoint) ||
-        !stationHostHealthMatches(h0, h1)
-      )
-        throw new Error("Station Host causal readiness evidence changed.");
-      return { endpoint: e1.endpoint, health: h1, holders, session };
-    } catch (error) {
-      session?.dispose();
-      if (admitted || isStationHostCompatibilityError(error) || ports.now() >= input.deadlineMs)
-        throw error;
-      const current = await ports.probeEndpoint(input.socketPath, input.deadlineMs);
-      if (
-        current.status !== "listening" ||
-        !stationHostEndpointsMatch(e0.endpoint, current.endpoint)
-      )
-        throw error;
-      await retryDelay(ports.now, input.deadlineMs);
-    }
-  }
-  throw new Error("Station Host causal readiness deadline exceeded.");
-}
-
-export type CausalStationHostStartResult =
-  | {
-      status: "transferred";
-      endpoint: StationHostExactEvidence["endpoint"];
-      health: HostHealthResult;
-      session: StationHostLifecycleSession;
-    }
-  | { status: "failed"; error: unknown; childDisposition: "not-spawned" | "settled" | "unproven" };
-
-/**
- * ADAPTER
- *
- * Starts one direct child and transfers it only after cutoff-bound causal proof.
- */
-export async function startCausalStationHost(
-  input: {
-    socketPath: string;
-    stateDir: string;
-    hostCommand: StationHostCommand;
-    detached: boolean;
-    expectedBuildVersion: string;
-    startupCutoffMs: number;
-    deadlineMs: number;
-    validate?: (session: StationHostLifecycleSession) => Promise<void>;
-  },
-  overrides: Partial<CausalStationHostEvidencePorts> & {
-    spawnHost?: (input: SpawnStationHostInput) => ChildProcessLike;
-  } = {},
-): Promise<CausalStationHostStartResult> {
-  const ports: CausalStationHostEvidencePorts = {
-    openSession: overrides.openSession ?? openStationHostLifecycleSession,
-    probeEndpoint: overrides.probeEndpoint ?? readStationHostEndpoint,
-    readHolders:
-      overrides.readHolders ??
-      ((path, deadlineMs) => readUnixSocketHolderPidsAsync(path, { deadlineMs })),
-    now: overrides.now ?? Date.now,
-  };
-  let child: ReturnType<typeof startStationHostProcess> | undefined;
-  let candidate: CausalCandidate | undefined;
-  try {
-    if (ports.now() >= input.startupCutoffMs)
-      throw new Error("Station Host startup cutoff was reached before spawn.");
-    child = startStationHostProcess(
-      {
-        argv: [...input.hostCommand, "--socket", input.socketPath, "--state-dir", input.stateDir],
-        spawnOptions: { detached: input.detached, stdio: "ignore" },
-      },
-      {
-        ...(overrides.spawnHost === undefined ? {} : { spawnHost: overrides.spawnHost }),
-        now: ports.now,
-      },
-    );
-    candidate = await readCausalCandidate(
-      {
-        socketPath: input.socketPath,
-        expectedBuildVersion: input.expectedBuildVersion,
-        deadlineMs: input.startupCutoffMs,
-        ...(input.validate === undefined ? {} : { validate: input.validate }),
-      },
-      ports,
-    );
-    if (!child.transfer(candidate.holders, input.startupCutoffMs))
-      throw new Error("Station Host child ownership was not proven.");
-    return {
-      status: "transferred",
-      endpoint: candidate.endpoint,
-      health: candidate.health,
-      session: candidate.session,
-    };
-  } catch (error) {
-    candidate?.session.dispose();
-    if (child === undefined) return { status: "failed", error, childDisposition: "not-spawned" };
-    const settled = await child.cleanup(input.deadlineMs).catch(() => false);
-    return { status: "failed", error, childDisposition: settled ? "settled" : "unproven" };
-  }
-}
-
 export function stationHostEndpointsMatch(
   left: StationHostExactEvidence["endpoint"],
   right: StationHostExactEvidence["endpoint"],
@@ -274,10 +115,5 @@ export function stationHostEvidenceMatches(
     stationHostHealthMatches(left.health, right.health) &&
     left.buildIdentity === right.buildIdentity &&
     JSON.stringify(left.terminals) === JSON.stringify(right.terminals)
-  );
-}
-function retryDelay(now: () => number, deadlineMs: number): Promise<void> {
-  return new Promise((resolve) =>
-    setTimeout(resolve, Math.max(0, Math.min(25, deadlineMs - now()))),
   );
 }
